Extract change batching to parent class
