+++ packages/database/src/changeHandlers/AnalyticsRefresher.js
@@ -4,71 +4,21 @@
  */
 
 import winston from 'winston';
+import { ChangeHandler } from './ChangeHandler';
 
-const REFRESH_DEBOUNCE_TIME = 1000; // wait 1 second after changes before refreshing, to avoid double-up
+export class AnalyticsRefresher extends ChangeHandler {
+  debounceTime = 1000;
 
-export class AnalyticsRefresher {
-  constructor(database, models, refreshDebounceTime = REFRESH_DEBOUNCE_TIME) {
-    this.database = database;
-    this.models = models;
-    this.refreshDebounceTime = refreshDebounceTime;
-    this.scheduledRefreshTimeout = null;
-    this.scheduledRefreshPromise = null;
-    this.scheduledRefreshPromiseResolve = null;
-    this.activeRefreshPromise = null;
-    this.changeHandlerCancellers = [];
+  constructor(models) {
+    super(models);
+
+    this.changeHandlers = {
+      answer: this.refreshAnalytics,
+      surveyResponse: this.refreshAnalytics,
+    };
   }
 
-  listenForChanges() {
-    this.changeHandlerCancellers = [
-      this.models.answer.addChangeHandler(this.handleSourceTableChange),
-      this.models.surveyResponse.addChangeHandler(this.handleSourceTableChange),
-    ];
-  }
-
-  stopListeningForChanges() {
-    this.changeHandlerCancellers.forEach(c => c());
-    this.changeHandlerCancellers = [];
-  }
-
-  handleSourceTableChange = () => {
-    return this.scheduleAnalyticsRefresh();
-  };
-
-  async scheduleAnalyticsRefresh() {
-    // wait for any active refresh to finish before scheduling a new one
-    await this.activeRefreshPromise;
-
-    // clear any previous scheduled rebuild, so that we debounce all changes in the same time period
-    if (this.scheduledRefreshTimeout) {
-      clearTimeout(this.scheduledRefreshTimeout);
-    }
-
-    if (!this.scheduledRefreshPromise) {
-      this.scheduledRefreshPromise = new Promise(resolve => {
-        this.scheduledRefreshPromiseResolve = resolve;
-      });
-    }
-
-    // schedule the rebuild to happen after an adequate period of debouncing
-    this.scheduledRefreshTimeout = setTimeout(() => {
-      this.activeRefreshPromise = this.refreshAnalytics();
-    }, this.refreshDebounceTime);
-    return this.scheduledRefreshPromise;
-  }
-
-  refreshAnalytics = async () => {
-    // remove timeout so any jobs added now get scheduled anew
-    const existingResolve = this.scheduledRefreshPromiseResolve;
-    this.scheduledRefreshTimeout = null;
-    this.scheduledRefreshPromise = null;
-
-    // get the subtrees to delete, then run the delete
-    await AnalyticsRefresher.executeRefresh(this.database);
-    existingResolve();
-  };
-
-  static async executeRefresh(database) {
+  static refreshAnalytics = async database => {
     try {
       const start = Date.now();
       await database.executeSql(`SELECT mv$refreshMaterializedView('analytics', 'public', true);`);
@@ -77,5 +27,7 @@ export class AnalyticsRefresher {
     } catch (error) {
       winston.error(`Analytics table refresh failed: ${error.message}`);
     }
-  }
+  };
+
+  refreshAnalytics = async () => AnalyticsRefresher.refreshAnalytics(this.models.database);
 }
