+++ packages/database/src/changeHandlers/ChangeHandler.js
@@ -1,0 +1,114 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+import assert from 'assert';
+
+export class ChangeHandler {
+  /**
+   * A map of change pushers by record type. Each pusher can return a list of changes that will be
+   * queued and scheduled for handling at a later stage, when changes stop coming through.
+   *
+   * If no pusher is defined for the record type, then the `changeDetails` generated by the db
+   * will be simply added to the change queue
+   *
+   * @protected
+   * @type {Record<string, Function>}
+   */
+  changePushers = {};
+
+  /**
+   * A map of change handlers by record type. Each handler will be used to process queued changes
+   * of that record type
+   *
+   * @protected
+   */
+  changeHandlers = {};
+
+  /**
+   * Wait 1 sec after changes before handling the change queue, to avoid double-up.
+   * Can override in child classes to fine-tune the timing of queue handling
+   *
+   * @protected
+   */
+  debounceTime = 1000; // ms
+
+  /**
+   * @type {Record<string, Function>} Changes by record type
+   */
+  changeQueue = {};
+
+  scheduledTimeout = null;
+
+  scheduledPromise = null;
+
+  scheduledPromiseResolve = null;
+
+  activePromise = null;
+
+  changeHandlerCancellers = [];
+
+  constructor(models) {
+    this.models = models;
+  }
+
+  setDebounceTime(debounceTime) {
+    this.debounceTime = debounceTime;
+  }
+
+  listenForChanges() {
+    this.changeHandlerCancellers = Object.keys(this.changeHandlers).map(recordType => {
+      this.changeQueue[recordType] = [];
+      const changePusher = this.changePushers[recordType] || (changeDetails => [changeDetails]);
+
+      const pushAndSchedule = async changeDetails => {
+        const newChanges = changePusher(changeDetails);
+        this.changeQueue[recordType].push(...newChanges);
+        return this.scheduleChangeQueueHandler(recordType);
+      };
+      return this.models[recordType].addChangeHandler(pushAndSchedule);
+    });
+  }
+
+  stopListeningForChanges() {
+    this.changeHandlerCancellers.forEach(c => c());
+    this.changeHandlerCancellers = [];
+  }
+
+  async scheduleChangeQueueHandler(recordType) {
+    // wait for any active refresh to finish before scheduling a new one
+    await this.activePromise;
+
+    // clear any previous scheduled rebuild, so that we debounce all changes in the same time period
+    if (this.scheduledTimeout) {
+      clearTimeout(this.scheduledTimeout);
+    }
+
+    if (!this.scheduledPromise) {
+      this.scheduledPromise = new Promise(resolve => {
+        this.scheduledPromiseResolve = resolve;
+      });
+    }
+
+    // schedule the rebuild to happen after an adequate period of debouncing
+    this.scheduledTimeout = setTimeout(() => {
+      this.activePromise = this.handleChangeQueue(recordType);
+    }, this.debounceTime);
+
+    return this.scheduledPromise;
+  }
+
+  handleChangeQueue = async recordType => {
+    // remove timeout so any changes added now get scheduled anew
+    this.scheduledTimeout = null;
+    this.scheduledPromise = null;
+
+    const changeQueueHandler = this.changeHandlers[recordType];
+    assert.ok(changeQueueHandler); // Change handler for the record type should be defined
+
+    await changeQueueHandler(this.changeQueue[recordType]);
+    this.changeQueue[recordType] = [];
+    this.scheduledPromiseResolve();
+  };
+}
